Version 0.6.0
So many changes that they can't all be listed.
The big ones are:
- Code cleanups and reorganization.
- Added Custom PlotUpgrades system
- Added .obj model file loader so plugins can now add their own models
to the game!
- Added support for .DDS image file loading, using .DDS textures GREATLY
decreases asset memory footprint on the harddisk(in many cases)
- Fixed many UI problems
- Plugin Settings/Store windows now look nicer
- Added more events such as: MainMenu_Loaded, Level_Loaded to the
SiscosHooks system so you don't need to create a custom script to detect
it.
- Added many new helper classes for things such as: Sounds, Textures,
HUD, ViewModel, GameTime(Pausing), and more!
- Changed the way plugins add thumbnails and icons for themselves, the
icon and thumbnail resource filename must now be specified in the
PLUGIN_DATA structure.